(refactor) nginx -> gunicorn via TCP rather than unix socket
Although this would be slightly less performant, what is currently

  nginx -> gunicorn via TCP over unix socket -> django via WSGI

will be replaced by

  nginx -> aiohttp proxy -> gunicorn via (?) -> django via WSGI
                         -> applications via HTTP over TCP
                         -> SSO via HTTP over TCP

And so it's simpler for all of the proxy's outgoing connections to be
via TCP. For example, would be able to use the same aiohttp client
session.

A future simplification may be to replace gunicorn with something else,
perhaps aiohttp-based. However, gunicorn -> django is a "tried and
tested" method, so keeping that for now.