(fix) Stop idle 'SPAWNING' Fargate containers
If a user has started the spawn process for an application, but not
viewed it, it will remain SPAWNING. The previous code would set the
model to STOPPED, so it disappears, but the Fargate container will still
be running.

The previous changes that made stopping the Fargate containers
asynchrounous makes this change robust against the container not
existing for whatever reason, such as some manual tinkering in the AWS
console.