(feat) Don't start the application from ajax requests
If the user has an application spawning in one tab, and stops in
another, then the spawning tab would have respawned the application
without any visual change to the user. For something like JupyterLab
that spawns for a long time, this would then result in the counter
reaching 0:00 until the application loaded.

In this change, the spawning tab would receive a 500, and then reload.
The application would attempt to respawn still, but at least the counter
would reset. Could refine this further, but KISS for now.