(feat) Bump datasets max connections
The datasets db is a Aurora PostgreSQL db.r5.large which supports up to
1000 connections. The value changed here is for a single application
instance, of which there are two currently.

Bumping since, unlike queries to the main admin db, queries to the
datasets dbs can be long lasting, e.g. to download tables with GBs of
data.

Not bumping crazy high to allow for scaling of the application
instances, and noting that connections to the datasets dbs are performed
from the tools