(fix) Admin report order, and display of null values
Order rows with runtime as those above those without.

I don't think there is a perfect way to deal with the lack of
information of runtime of existing applications, but without this, rows
that have NULL as their total runtime are always above those that have
some values. Suspect there will still need to be tweaks, but this change
gives a more reasonable order with the minimum of code.