(feat) Own RStudio image
The Rocker image does some user/permissions tinkering on startup, which
I suspect causes a race condition with the s3sync sidecar container that
means when it attempts to write to the home directory it doesn't have
permissions to do so, and exits.

We could have put some logic in there to retry. However, managing our
own image is probably good in the long run, for example to minimise the
image size.