(feat) Separate JupyterLab Python and R
There is duplication. However, it is suspected that the JupyterLab R
version will be deemed unnecessary and removed, so not doing and
refactoring work to address this, as well as the JupyterLab mixed
Python/R image.

The new application types are hidden from the front page to a) not cause
issue during deployment, and b) be able to test them in production
before exposing them to users.