(feat) Application templates can be hidden via environment variable
This is getting ready for the JupyterLab Python and R split. The old
JupyterLab application template will be kept in the database for stats
reasons, and to not break running containers, but will be marked as not
visible, so it will be hidden from the front page and the tools page.

This will still allow non-visible applications to run by going to their
URL. After the split, it is expected that this will be then blocked.