Move python requirements from Dockerfile to requirements.txt
Currently we list all python dependencies as arguments to pip install
in our main Dockerfile. This works well enough for our short list
of packages, but has a number of downsides:

* It's not generally supported by most python tools. For example pyup,
  pip-tools and Github vulnerability alerts wouldn't normally pick it
  up for scanning.
* It requires maintaining a separate copy of the dependencies list in
  order to install dependencies locally (e.g. for running unit tests).

This moves the list of packages to a requirements.txt file, which is
copied to the docker container before the `pip install` step is run.

Because we're using alpine as a base image however, some packages still
need to be installed as alpine packages (psycopg2-binary for example
doesn't support alpine and we'd like to avoid building the package from
source each time), which makes requirements.txt currently incomplete.