Add a dataset eligibility criteria DB field
Makes it possible to store a list of access requirements for each
dataset.

The criteria are (potentially) different between each dataset, so
for now we're storing them as a list of strings (as opposed to a
many-to-many relationship with a new table).

This is using a postgres ArrayField to store data since it's the
easiest type to setup with Django admin and fits the data quite
well. One possible alternative would be a single text field, but
it would require us to use a special character (a newline for example)
as a delimiter between different criteria in order to render them
as bullet points.