Add dataset eligibility criteria field to the admin page
Allows setting and changed the access requirements for a dataset.

This is using the django-better-admin-arrayfield to render a dynamic
number of text inputs (with a "-" button next to each one and "Add
another" at the bottom).

One alternative I've looked at is using django.contrib SplitArrayField,
which renders a static number of inputs. While a fixed number isn't
an issue (we don't expect there to be more than 5 requirements),
the built-in widget puts all inputs on the same line and doesn't allow
removing an item from the middle of the list (you can remove text,
but you'd need to also move other requirements so that there's no gap).