Persist local postgres data when containers are recreated
By default, docker creates anonymous volumes for the postgres container.
This volume can be reused when `docker-compose up` is run, but would
normally be lost on `docker-compose down` or if a container is rebuilt,
taking all postgres data with it.

Creating a named volume makes sure the populated DB persists between
image/container changes.

The data can still be reset with `docker-compose down -v`.