Don't map host ports when running containers for test
`docker-compose run` won't apply published port mappings for the
target service, but it still attempts to launch postgres and redis
as is, which fails if there's another instance of either already
running on the same default port.

We don't need to open ports for test containers to the host, so we
can remove the mapping altogether, which still allows tests to talk
to the DB but shouldn't conflict with any other running containers.