Add static analysis and autoformatting checks to CI
Similar to uktrade/data-flow#8 this adds a make target to check and
autocorrect code formatting.

The steps run in CircleCI and can be run locally using `make check`
and `make format`. They require requirements-dev.txt packages to be
installed.

They checks aren't run inside the test docker container since it only
contains dataworkspace/ code, which would exclude other python files
from being checked.

This is also not adding mypy at the moment. Because our images are
based on alpine mypy has to be built from source in the test container,
which either complicates the build process or means we need to install
it separately, without listing it in the dev requirements file.

mypy also currently flags a lot of missing annotations, which would need
to be corrected first, so I'm leaving it for a separate task for now.