Add a model method to create a duplicate of a dataset
Using the recommended Django approach of duplicating a model instance
this creates a coy of the dataset and any related tables, view, links
etc.

There are a few packages available that add a general clone model mixin,
but none appear to be widely used and since a general case is much more
complex I'm going with a solution that hard-codes copied related objects
for now.

https://docs.djangoproject.com/en/2.2/topics/db/queries/#copying-model-instances