Display list of data fields on the dataset page
Adds a preview of the dataset fields to help users make an informed
decision about requesting access.

The field list is derived from the dataset tables, view or custom query
and so should match the columns in a download CSV.

For master datasets containing one or multiple tables the table columns
are prefixed with the table name.

The fields are currently only visible to admin users and the section is
skipped if it failed to retrieve columns from the datasets DB.