Add a Django forms widget for a Design System checkboxes component
We need to render the search filters as a collection of checkboxes.
Normally we'd add the HTML directly to the template, but in this case
it involves a lot of templating logic for setting correct checkbox ids,
labels and setting the checked value based on the provided form data.

Django forms already handles a lot of this, but the default templates
for CheckboxSelectMultiple have a different layout than the GOV.UK
Design System component, so we need to write our own custom widget and
related templates.

Apart from custom classes and div structure, one big difference in
Django and GOV.UK form layout if that Django considers the form field
label to be outside the input template. The label is rendered when the
whole form is rendered, but not when an individual field is (for example
using `{{ form.field  }}`). In Design System however the field heading
is inside the field's div and fieldset tags, which should be rendered by
`{{ form.field }}`.

One way around this would be to use grouped lists of `choices` and render
the name of the group as the heading for the set of checkbox inputs.
The problem is that the grouped choices aren't supported out of the box
by `ModelChoiceField`, so we'd need to write additional logic to group
queryset results.

Instead, we add a new `group_label` argument to the widget constructor,
which should contain the desired heading for the field and is passed to
the template context.