Add a datasets search page with filters
Adds a new /datasets/ page with search and filters across datasets,
data cuts and reference datasets.

Search and filters are validated by a Django form and applied to
a dataset and (if required) reference dataset queries. The queries
are converted to dictionaries of common fields, which makes it
possible to combine and sort them with a `.union`, allowing us to
search across two separate models.

The `use` filter controls which types of datasets should be included,
with a special "0" value for reference datasets (to keep the same type
as the DataSet.type while avoiding clashing with potential future types).

This change also includes CSS rules for the search field, which
currently doesn't have a component in the GOV.UK Design System
(https://github.com/alphagov/govuk-design-system-backlog/issues/132).

The rules are based on the GOV.UK search box.

Things that are currently missing:

* Individual pagination links, currently there's only `Next` and
  `Previous`
* Refreshing the results when a filter is selected/deselected without
  clicking the search button
* Search box default text label, which requires JavaScript to hide the
  text if there's user input in the field