Redirect to the search home page if search form is not valid
Since there are no required fields, the only way for the search form
to not pass validation is if the query parameters have been modified
by hand. Since this is not a normal use case, we can drop the query
and redirect back to the search start page without trying to display
form error messages.