Update search results on filter / search input change
This adds an AJAX request to update dataset search results each time
a filter checkbox is selected/deselected or search keywords are modified.

Any click on a checkbox within the search form as well as a clicking
away from a search input after modifying the keywords will result in
an AJAX request, fade out the current search results and then replace
them with the new ones. This also resets the pagination, so new results
always start on page 1.

URL and browser history are updated on each click.

Previous results are cached in-browser to reduce the number of requests
as well as make history navigation faster.

This is mostly based on [Digital Marketplace implementation][dmsearch],
with some changes:

* Our search form is outside the results wrapper, so we pass in two
  selectors instead of one.
* Instead of having a separate API endpoint for returning separate
  HTML strings for each part of the page that needs to be updated,
  we request the whole search response and then replace contents
  of our search results by finding the corresponding element in the
  new response. This saves some development work on the Django side
  and since the majority of the page rendering is dealing with the
  results themselves shouldn't affect performance.
* All references to GOVUK toolkit are removed since we're using it
  as a stand-alone script instead of a plugin.

[dmsearch]: https://github.com/alphagov/digitalmarketplace-frontend-toolkit/blob/master/toolkit/javascripts/live-search.js