(feat) Sync .ssh directory from tools
For the upcoming connection via SSH from tools to gitlab, this will be
via a public/private key pair. It would be expected that you wouldn't
have to regenerate this every time a tool starts.

Note that for now, mobius3 does not store the mode of files, so there is
likely work needed there: otherwise SSH will complain about the private
key being too open.