(feat) Upload pack files on create
It looks like no IN_CLOSE_WRITE event is received for .pack or .idx
files. I suspect they are created via

- a separate temporary file, with name like tmp_pack_hBV4Alz, being
  created, modified and, closed;
- a _hard_ link is created to this file, with the final .pack name;
- the original tmp_pack_hBV4Alz name is deleted.

So after a repository is cloned, because there is no IN_CLOSE_WRITE
event, the upload of these files is never queued, and after a few
minutes, the sync notices these files are not in the bucket, and thinks
that these files must have been removed, and then decides to delete them
locally.

This change is somewhat hacky/brittle. If git changes the location of
these files for example, this fix would break.