(feat) Give more memory and CPU to visualisations
We could have some sort of per-visualisation configuration, but for now,
probably not worth it.

Bumping to the same amount of memory/CPU as the tools, which is 1 CPU,
and the maximum amount of memory for that CPU, 8GB.

I suspect for a lot of visualisations, such as one that has just been
setup for testing, their data will just be loaded into memory. Not
bumping memory higher for now to have a bit of visibilty on the sorts of
amounts that are needed.