Add a dev docker-compose file with inline start script
This adds a docker compose file specifically for local development.

The main differences are:
* It doesn't run celery
* It runs django dev server instead of gunicorn to enable hot reload
* it doesn't run nginx or pre-compile static files so changes in
  assets are applied without having to restart the container
* dataworkspace/ is mapped as a volume, so file changes are picked up

This inlines the updated start-dev.sh script into the compose file
itself to make it a bit easier to see what's going on.

This also removes the unused local Mac setup script since it's never
been finished and we've largely settled on using Docker containers
for local development.