Build dataworkspace docker container in Jenkins
Previously, our images were build by quay.io using a Github trigger
and we needed to wait for the build to finish before triggering a
release job in Jenkins or it would pick up an old master image tag.

This adds a step to the Jenkins pipeline that builds a docker image
for dataworkspace admin and pushes it to quay before proceeding with
the release to dev. This way, we can kick off the release pipeline
automatically after merge and the correct image, tagged with commit
SHA will get released. This also allows us to resume older pipelines
paused before staging/prod since they'll be tied to a unique immutable
tag.

This merges existing prepare step and a k8s agent build step into one,
since both require a workspace checkout. The agent is defined within
the stage, allowing pipeline itself to use `agent none` and avoid
blocking an executor while waiting for manual approval.

The build step is locked, preventing simultaneous builds, but making
it safe to merge things in quick succession without causing a release
race.