Run ecs-pipeline jobs directly
This removes the intermediate data-workspace-{env} jobs and runs
ecs-pipeline with required parameters directly.

This reduces the number of simultaneous jobs running in Jenkins during
release and makes it easier to get to the logs in case of failure.

One of the benefits of a single release pipeline is that it shows
new commits since last job run, making it easier to tell what changes
are about to go live. This benefit is reduced if some releases are
going through separate jobs, so the idea is to remove individual env
release jobs once this is merged. In case we ever need a quick hotfix
release to prod while dev/stage pipeline is blocked/broken we can
always use ecs-pipeline job directly.

In addition, this adds the locks around release jobs making sure
simultaneous merges don't overtake each other between stages.