Add user access editing to dataset admin page
This allows viewing / modifying users that have access to a dataset
from the dataset admin page, similar to how a list of dataset
permissions for each user can be edited on the user page.

If the permissions are modified, an admin log event is written. The
event is associated with the user object, since that's how we record
permission changes currently.

This only applies to master datasets and data cuts, since reference
data currently doesn't require access requests.