(feat) Start of Visualisations page
- Fetches a list of Visualisation projects from GitLab on a new page
- This page is accessible to superusers and users who have the
  develop_visualisations permission
- There is a new top level tab, visible only to those how have
  develop_visualisations explicity, so superusers can maintain
  the view of most users as they wish.

This is is the first of a sequence of changes on this feature. Right now
this isn't that useful, but wanting feedback from production, e.g. for
speed of page loads that depend on synchronous responses from GitLab,
and feedback from other developers on permissions/code etc. Am hoping we
can live without having to do some ajax-y requests from the browser, and
without server-side caching, but we will see.