Migrate app CSP from proxy to Django
In order to enable nonces for CSP, to support GTM, we need to be able to
pass the nonce into the CSP headers via Django. At the moment CSP is set
by the proxy.

This patch moves ownership of CSP for Django pages from the proxy to the
app. The CSP should be identical other than the addition of the nonce.

Ticket: https://trello.com/c/C1xJiFLv/880