Split test databases into dataworkspace and datasets
Up till now, our dev and test environments were using the same
`postgres` database for both Data Workspace tables and as the
datasets DB.

This made it easier to setup an environment and makes it slightly
easier to create test data since we can use Data Workspace tables
as datasets, which some of the application tests do.

There is a number of downsides though:

* Using `postgres` makes it hard to tell whether we're doing set up
  and teardown correctly since it always exists.
* Using a single DB is different from what we're doing in every other
  environment, so can potentially hide bugs
* Django test set up considers DATABASE records with the same name to
  be aliases, so the setup/teardown process is different and appears
  to be less predictable

Splitting the databases requires updating some of the existing tests
and making sure our local postgres container creates both on startup.

This modifies the postgres container start up script to create
everything needed for application tests and local application dev in
one Dockerfile, since we mostly want the same things in both.

This also fixes the application tests that were previously using the
auth_user table as a dataset to use the test_dataset fixture set up
in postgres start up script.