Combine dataworkspace Dockerfile into multi-stage build
Our test docker image was being built on top of the live one using
a separate Dockerfile. This created two issues:

* To build an up-to-date test image the base image needs to be built
  with a separate command, which is easy to miss when not using the
  make target.
* Since base image copies the dataworkspace/ files any change to them
  would invalidate the cache for all test image layers, requiring all
  dev requirements to be reinstalled during the test run. This adds an
  additional 3-4 minutes to the test runtime.

This merges Dockerfile with Dockerfile-test into a 3-stage build
process, separating the base image with common dependencies from
live and test images. By default, when running `docker build` a
live image will be built and tagged. When running a build with a
`test` target, a test image is created, which installs test
dependencies and copies application files on top of the common base.

The main downside of this approach is that since there's currently no
way to skip a build stage, building a live image will also build the
test one, increasing Jenkins build times.