Loosen up CSP on reference dataset admin page
We use CKEditor to render rich text fields on the add/change admin pages
for reference datasets. Our latest CSP update was too strict, leading to
the text field no longer working as an inline script injected by CK
Editor is no longer executed by the browser.

This script is dynamically built by the ckeditor javascript and it
doesn't appear possible to update the CSP headers with a hash of the
script contents or pass in our CSP nonce, so we will have to settle with
downgrading the CSP security on this page for the time being.