(feat) Show Tools in admin
There's going to have to be some changes regarding how host work for
visualisations. Right now, they're a strange(?) host_exact/host_pattern
combo, mixed in with some config and quite spread around logic: the
feature-to-complexity ratio is not good.

I suspect for all applications (tools _and_ visualisations) will just
have a host_prefix field (that will have to be unique) that will
determine which application a given requests maps to.

To get there, will have to make some changes to Tools config, so making
that possible in admin. It's not particularly friendly to have all the
Tools config done all by environment variables, so moving away from that
pattern (as much as possible at least).