(feat) Remove unused 'user-provided' app
The "UserProvided" application template was to be some sort of
template-template for visualisations. This isn't how they currently run,
and not the plan on how they will run. Each is a separate
ApplicationTemplate. For visualisations, this will have to be created
just from code, not from code and an existing ApplicationTemplate

[I suspect each ApplicationTemplate will have to refer to a
SpawnerOptions model to know what options to pass to the Spawner, and
these SpawnerOptions may need to be created from environment variables,
and there be an environment variable that determines what SpawnerOptions
are used for visualisations by default.]