(refactor) Un-generalise public_host -> docker tag for visualisations
For visualisations, there are two slightly unnecessary generalisations
to convert the host to a docker tag:

- Application.host_pattern that converted the host into some arbitrary
  dictionary

- Application.spawner_options.CONTAINER_TAG that then converted from
  this dictionary to a docker tag

However for visualisations the only transformation this ever made was to
convert the exact value of public_host to the docker tag. There isn't a
plan to change this, and even if there were, I suspect will want to do
this from code for all visualisations, so it will be awkward to do this
for all visualisations if it involved their config.