Add code snippets for master datasets
We want users to be able to quickly get going with datasets in tools on
the Data Workspace, and one way we can do this is to help lower the
barrier to entry. One way to do this is to provide the boilerplate code
that they need to start making queries against the data.

This patch provides code snippets for master datasets that can be copied
(via a single click) and pasted into a tool, to immediately start
browsing the dataset. This "copy code" snippet is heavily reused from
the GOV.UK Design System, included their main dependency Highlight.JS,
which we vendor here as we don't have NPM infrastructure set up to pull
it in.

Ticket: https://trello.com/c/CrpejTJJ/558