Actually de-duplicate datasets properly
I tried to fix this before in
https://github.com/uktrade/data-workspace/pull/530 but hadn't understood
the problem properly.

When a user filters for access as well as use (but not reference
datasets), then we get multiple instances of each dataset returned if
that dataset doesn't require authorisation but has multiple people
specifically authorised to use it.

This fix adds proper querying on public datasets to ensure that only one
instance is returned, and also changes the context passed into templates
to a list of dicts, rather than a list of dataset objects, which is
consistent with what's passed in when "reference" datasets are filtered
for.