Remove intermediary queue from CSV data downloads
When a user downloads a CSV we used to spawn a gevent greenlet that
would fetch results from DB and put them in a queue. Django then
returned a streaming response that was reading data from that queue
to return to the client. This allowed us to prefetch results, making
downloads potentially faster on quick connections since the query is
executed in parallel with the request being streamed.

This approach however turned out to be prone to creating hard to
debug problems. It used to cause incomplete downloads on slow
connections without reporting the error to the user until we tweaked
queue timeouts and linked greenlet error handling together. Since
the queue still required a timeout it was still possible for a very
slow connection to fail an otherwise successful download if it took
too long to fetch data.

Another recent issue potentially caused by this is that if a client
has disconnected a spawned greenlet was left behind with a stuck DB
transaction. That transaction would idle for hours, locking DB tables
and preventing data-flow from swapping dataset tables.

To avoid these and to potentially make this code easier to reason
about we remove the queue and stream DB response directly to WSGI.
We still use `.fetchmany` to `yield` things in chunks as yielding
each row slowed down the response (psycopg2 still uses itersize to
fetch results in bulk, but WSGI server has to call the stream
function a lot more)