Upgrade packages to fix build
The current dockerfile fails to build due to a number of dependencies
that have lately become unavailable (git and git-man). There's also been
a problem installing our pinned conda packages, whereby it's unable to
find "/opt/conda/bin/python3.7". Upgrading to the latest version of
conda fixes this - but it's not obvious what exactly changed since the
time that the image could successfully build (mid-Feb being the last
time this dockerfile was updated).