(fix) 'Skip to main content' link
The base href is set to `/` on the root domain, partially to make links
on the spawning and error pages, which are on other domains, work as
expected. However, this means that links with just a #fragment will
always go to the #fragment on the root path of the root domain, unless
prefixed by the path of the current page.