Upgrade JupyterLab to v2.0.0
Re-work existing database access plugin to continue working with the new
JupyterLab extension API.

Notes:
* We update git/git-man/sudo apt packages as the old ones are no longer
available, looks like removed due to vulnerabilities.
* Remove `-s` option from `conda clean` as it's been deprecated. Not
replaced it with anything as it's there to clean up source files from the
image build, but the new process is `conda build --clean-sources`, which
isn't available by default, i.e. requires installing a new package.
Doesn't feel necessary to do.
* We allow `blob:` in img-src for CSP as it seems matplotlib now manages
these as blobs rather than as `data:`.
* Add custom styling to the database access button for the JupyterLab
plugin. As part of the upgrade, icons were reworked and it became
transparent due to jupyterlab styling. This style simply overrides it and
sets the icon colour to black again. A bit of a hack, but didn't feel
worth spending much more time on it.