Use gunicorn for local dev server
We were using Django test server when running Data Workspace locally,
but it has some differences in how it behaves compared to production,
for example in the way it handles interrupted chunked responses for
data downloads (gunicorn correctly fails the response while Django
doesn't).

To save some time debugging these in the future this switches dev
server to use gunicorn with auto-reload. This should behave the
same way django test server does when python/html or asset files
are changed on disk.