(feat) Access mirrors from apps
Added slightly magic /__mirror/ path accessible on the domain of each
application that routes to our S3 mirror. Specifically for applications
that need map tiles, which are too big to put into each application.

The /__mirror/ path is behind SSO and the VPN, but not checking if the
use has access to the application, or even if it exists to keep latency
at minimum and not touch the DB.