Instrument Django-Debug-Toolbar, if installed
The debug toolbar adds some nice profiling info (e.g. sql queries for
the page, load time, headers/settings, etc) which can be handy in a dev
environment.

We only load the middleware if we're in debug mode and the library is
available, and we only add the library to the dev requirements, so it
should never be available in a production (even dev/staging)
environment.