Visualisations catalogue item model
Add a model to the application for visualisations in the search
catalogue, including an admin interface. The model is one-to-one linked
to a visualisation application template, which contains the launch
configuration for the vis. Some of the information is kept in sync
between visualisation templates and the catalogue item (at the moment
just the name).