(refactor) Reduce usage of application 'name', to remove later
This is so there will be less to setup for each visualisation.

The change of the database role _should_ be safe no matter the
characters in the host_basename: it gets escaped using sql.Identifier
before being sent to the database.

The usage of `name` in views.py appears to be unused in the template, so
it doesn't need to be replaced by anything.