(feat) Use visualisation topic rather than a group to search visualisations
This removes the requirement for an admin to move projects into the
visualisations group, since users can add topics to their own projects.

I realise that the endpoint is paginated, and the query does not follow
the pagination. However, it already wasn't, so this change doesn't go
backwards in this respect. In fact, mitigating the issue for now by
bumping the page size from the default 20 to 100. Following the
pagination is left until later.

It does seem that there isn't a way to query for projects with a
specific topic: this has to be done as a manual filter of the results.
However, since we can use min_access_level for the current user, I think
this is still acceptable.