(feat) Success alert more similar to error alert
I realised it was a bit odd when a field in the form is autofocused: the
alert is skipped and page scrolls down past it.

So now (all in common with the GDS error alert):

- It has role="alert"

- It autofocuses on page start (which means later elements in the page
  are _not_ autofocused)

- It's _not_ in the keyboard tab order of the page,  i.e. it is only
  focused on page load, or click from what I can tell. This is a bit odd
  to me, since it doesn't seem that getting the focus back on the element
  is possible with just keyboard navigation, but it is consistent with the
  GDS error alert.

- It has a yellow outline when focused [the yellow==focused is in common
  with GDS, although depending on element, sometimes the yellow is an
  outline, and sometimes a background]

There are lots of options + discussion at
https://github.com/alphagov/govuk-design-system-backlog/issues/2. So I
suspect there could be larger changes to the alert later [deliberately
not doing these now].

Potentially clients could have cached the old CSS file for up to a week.
Then on the alert they would see the browser's default outline of a
focused element: in Chrome on Mac OS this is a light blue outline.
Treating this as the least bad of available options... e.g. bumping the
file name of the CSS file to cache bust it I realise may be even worse,
since during a deployment, requests go between the old and new
instances, so potentially a client can see the HTML that contains the
file name of a CSS file, but then the request for that CSS file goes to
an instance without it, the browser gets a 404 for the CSS file, and
they will see a horribly broken site.