Add a Data Vis UI catalogue edit page
Add a page to the data visualisation UI that allows users to edit the
information that will be displayed on the catalogue page (i.e. this is
functionally-equivalent to the Django admin page).

Some things of note:
* Descriptions for reference/datacut/master datasets use a richtext
field editor for descriptions. This requires blowing a hole in our CSP
header, and generally provides features that aren't needed or desirable
for our content creators. The visualisation description therefore
supports basic markup - paragraphs and lists only. We can extend this
further if we need additional work in the future.
* We'll probably want to do some work around simplifying how we render
forms directly using GOV.UK Design System styles - at the moment it's
quite manual and fiddly, including specifying classes in Python code.

Ticket: https://trello.com/c/CvOWjmvD/978