Add "purpose" to the search result descriptions
Adds the filter name label based on the dataset type to each result.

We add the type to the reference datasets by making Django ORM generate
a `SELECT ..., 0 as "type"` query.

Django orders model fields before any additional fields like annotations
in generated SQL statement which means doing a UNION between a model
field in one query and an extra field in another results in different
columns getting connected into a single result one (which sometimes
leads to a type error and sometimes succeeds with values ending up in
the wrong place). To avoid this, we alias model field `type` with
`.annotate` in the datasets query, making sure it's added to the end
of SELECT field list.