(feat) Lock schema for all permissions changes on that schema
Potentially this could be more fine-grained: however revoking performs
actions on all tables in a schema, so would probably have to make that
per table to have more fine grained locking.

Hopefully the fact that the locks are held for one grant makes it that
other requests can proceed if possible.