Automatically grant vis creators the management perm
We have a permission that manages user access to unpublished catalogue
items. This patch automatically gives the permission to visualisation
creators as they browse the visualisation UI,by piggy-backing on
`gitlab_has_developer_access`, which is used to check whether a user has
developer+ access to a given visualisation.

The django-side permissions is global, so they are granted the ability
to see *all* unpublished visualisations. This is fine - and possibly
even preferrable - so that creators can see/crib from other people's
work.

Ticket: https://trello.com/c/0ZvQ3jK4/1012