(feat) Avoid surfacing IntegrityError when starting an application
If two parallel requests are received to start an application, the
second of these would error. Instead, we silently "succeed" by returning
the response as though the user succesfully started to start the
application.

This is likely to be more likely as visualisations get more popular,
since multiple people can start the same visualisation at the same time.