Send an email notification to visualisation contacts on access request
Sends an email through GOV.UK Notify to primary and secondary contacts
with the info about visualisation access requests.

The email contains a link to the "give access" page with the user
email address and related Zendesk ticket ID in an encrypted token.

This allows us to pre-populate the email field and update the ticket
in Zendesk if the visualisation owner grants access.

We're encrypting the token to avoid leaking the email in web server
logs. Another approach would be to send the user ID with HMAC (to
avoid exposing user emails by modifying the ID), but encryption
achieves the same result without having to look up the user object
on the "give access" page.