Pre-fill user email when giving access by following an email link
Pre-fills requester email address on "give access" page when the
visualisation developer clicks on the email link.

If the access is granted this adds a comment to the related Zendesk
ticket and marks it as solved, but only if the access is given to the
same email address listed in the original request.